Fix refresh content pool + Refactoring
