+++ src/components/UserActions/UserActions.jsx
@@ -4,28 +4,33 @@ import Deposit from './Deposit/Deposit';
 import Redeem from './Redeem/Redeem';
 import Repay from './Repay/Repay';
 
-function UserActions({ account, setStateStale, stateStale }) {
+// TODO refactoring
+function UserActions({ account, setStateStale, stateStale, updateData }) {
 	return (
 		<div>
 			<Deposit
 				account={account}
 				setStateStale={setStateStale}
 				stateStale={stateStale}
+				updateData={updateData}
 			/>
 			<Redeem
 				account={account}
 				setStateStale={setStateStale}
 				stateStale={stateStale}
+				updateData={updateData}
 			/>
 			<Borrow
 				account={account}
 				setStateStale={setStateStale}
 				stateStale={stateStale}
+				updateData={updateData}
 			/>
 			<Repay
 				account={account}
 				setStateStale={setStateStale}
 				stateStale={stateStale}
+				updateData={updateData}
 			/>
 		</div>
 	);
